backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
groq==0.11.0
httpx==0.27.0
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
pypdf==3.17.4
sentence-transformers==2.3.1
pydantic==2.5.3
python-multipart==0.0.6
pymongo==4.6.1
tiktoken==0.5.2